Move inbox id methods into separate mod
