+++ bindings_node/src/mls_client.rs
@@ -12,13 +12,9 @@ use tracing_subscriber::{filter::EnvFilter, fmt, prelude::*};
 pub use xmtp_api_grpc::grpc_api_helper::Client as TonicApiClient;
 use xmtp_cryptography::signature::ed25519_public_key_to_address;
 use xmtp_id::associations::builder::SignatureRequest;
-use xmtp_id::associations::generate_inbox_id as xmtp_id_generate_inbox_id;
 use xmtp_id::associations::unverified::UnverifiedSignature;
-use xmtp_mls::api::ApiClientWrapper;
 use xmtp_mls::builder::ClientBuilder;
 use xmtp_mls::identity::IdentityStrategy;
-use xmtp_mls::retry::Retry;
-use xmtp_mls::storage::consent_record::StoredConsentRecord;
 use xmtp_mls::storage::{EncryptedMessageStore, EncryptionKey, StorageOption};
 use xmtp_mls::Client as MlsClient;
 
@@ -125,36 +121,6 @@ pub async fn create_client(
     account_address,
     signature_requests: Arc::new(Mutex::new(HashMap::new())),
   })
-}
-
-#[napi]
-pub async fn get_inbox_id_for_address(
-  host: String,
-  is_secure: bool,
-  account_address: String,
-) -> Result<Option<String>> {
-  let account_address = account_address.to_lowercase();
-  let api_client = ApiClientWrapper::new(
-    TonicApiClient::create(host.clone(), is_secure)
-      .await
-      .map_err(ErrorWrapper::from)?,
-    Retry::default(),
-  );
-
-  let results = api_client
-    .get_inbox_ids(vec![account_address.clone()])
-    .await
-    .map_err(ErrorWrapper::from)?;
-
-  Ok(results.get(&account_address).cloned())
-}
-
-#[napi]
-pub fn generate_inbox_id(account_address: String) -> String {
-  let account_address = account_address.to_lowercase();
-  // ensure that the nonce is always 1 for now since this will only be used for the
-  // create_client function above, which also has a hard-coded nonce of 1
-  xmtp_id_generate_inbox_id(&account_address, &1)
 }
 
 #[napi]
